+++ deepparse/cli/__init__.py
@@ -1,3 +1,4 @@
 # pylint: disable=wildcard-import
-from .parsing import *
-from download import *
+from .parse import *
+from .download import *
+from .tools import *