+++ examples/parse_addresses.py
@@ -15,7 +15,9 @@ test_dataset_name = "predict"
 download_from_url(test_dataset_name, saving_dir, file_extension=file_extension)
 
 #  Now let's load the dataset using one of our dataset container
-addresses_to_parse = CSVDatasetContainer("./a_path.csv", column_names=["address_column_name"])
+addresses_to_parse = CSVDatasetContainer(
+    "./a_path.csv", column_names=["address_column_name"], is_training_container=False
+)
 
 # We can sneak peek some addresses
 print(addresses_to_parse[:2])